Refactor Database module and Relation iterator
- Split `relvar-core/src/database.rs` into `relvar-core/src/database/mod.rs`, `relvar-core/src/database/constraint_manager.rs`, and `relvar-core/src/database/error.rs`.
- Extracted `ConstraintManager` and `DatabaseError` to their own modules to reduce "God Object" complexity.
- Implemented `IntoIterator` for `Relation` to allow consuming tuples.
- Refactored `Database::update` and `Database::delete` to use `into_iter()` on relations, avoiding unnecessary tuple cloning.
- Verified all changes with existing tests and added a new test for `Relation::into_iter`.
